jeg har laget flere python filer of sortert ut funksjonene fra Tours inn i flere filer for mer oversikt
