+++ src/backend/autentication/user_online.json
@@ -1,1 +1,1 @@
-{"user_online": ""}
+{"user_online": "esel"}
